Remove `http_responder` subsystem
